rearranging test cases to follow refactor. Oddly, error message tests are failing (timeout state not updating in test, but message is)
