+++ src/tests/utlilities.test.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import { shallow, mount } from 'enzyme';
+import {getTermCourseList} from '../common/utilities';
+
+const seasonCourses = {
+    'GEOG': [{id: 0, name: "course 0"}, {id: 1, name: "course 1"}],
+    'MATH': [],
+    'LIFESCI': [{id: 3, name: "course 3"}],
+};
+const expectedSeasonCourses = [{id: 0, name: "course 0"}, {id: 1, name: "course 1"}, {id: 3, name: "course 3"}];
+
+describe('getTermCourseList', () => {
+  it('should return a list of all courses given a list of course types', () => {
+      // console.log(getTermCourseList(seasonCourses));
+    expect(getTermCourseList(seasonCourses)).toEqual(expectedSeasonCourses);
+  });
+});
